Validates recharge invoice data

+++ src/Payments.Mvc/Services/InvoiceService.cs
@@ -16,11 +16,13 @@ namespace Payments.Mvc.Services
     {
         private readonly ApplicationDbContext _dbContext;
         private readonly IEmailService _emailService;
+        private readonly IAggieEnterpriseService _aggieEnterpriseService;
 
-        public InvoiceService(ApplicationDbContext dbContext, IEmailService emailService)
+        public InvoiceService(ApplicationDbContext dbContext, IEmailService emailService, IAggieEnterpriseService aggieEnterpriseService)
         {
             _dbContext = dbContext;
             _emailService = emailService;
+            _aggieEnterpriseService = aggieEnterpriseService;
         }
 
         public async Task<IReadOnlyList<Invoice>> CreateInvoices(CreateInvoiceModel model, Team team)
@@ -40,6 +42,7 @@ namespace Payments.Mvc.Services
                 throw new ArgumentException("Account is inactive.", nameof(model.AccountId));
             }
 
+
             if (model.Type == Invoice.InvoiceTypes.Recharge)
             {
                 //TODO: Server side validation.
@@ -47,6 +50,41 @@ namespace Payments.Mvc.Services
                 //All recharge accounts must be valid.
                 //All credit recharge account must be 100% of total.
                 //If any debit recharge accounts, they must all be 100% of total.
+
+                if(model.RechargeAccounts == null || !model.RechargeAccounts.Any(a => a.Direction == RechargeAccount.CreditDebit.Credit))
+                {
+                    throw new ArgumentException("At least one credit recharge account is required for recharge invoices.", nameof(model.RechargeAccounts));
+                }
+
+                if(model.RechargeAccounts.Where(a => a.Amount <=0.0m ).Any())
+                {
+                    throw new ArgumentException("Recharge account amounts must be greater than 0.", nameof(model.RechargeAccounts));
+                }
+
+                foreach(var ra in model.RechargeAccounts)
+                {
+                    var validationModel = await _aggieEnterpriseService.IsRechargeAccountValid(ra.FinancialSegmentString, ra.Direction);
+                    if(!validationModel.IsValid)
+                    {
+                        throw new ArgumentException($"Recharge account '{ra.FinancialSegmentString}' is not valid");
+                    }
+                    // Ok, this could be called from an API, so we need to replace and chart strings that may get changed.
+                    if(validationModel.ChartString != ra.FinancialSegmentString)
+                    {
+                        ra.FinancialSegmentString = validationModel.ChartString;
+                        //log it?
+                    }
+                }
+
+                if(model.CouponId > 0)
+                {
+                    throw new ArgumentException("Coupons are not allowed for recharge invoices.", nameof(model.CouponId));
+                }
+
+                if(model.TaxPercent > 0.0m)
+                {
+                    throw new ArgumentException("Tax is not allowed for recharge invoices.", nameof(model.TaxPercent));
+                }
             }
 
             // find coupon
@@ -56,7 +94,6 @@ namespace Payments.Mvc.Services
                 coupon = await _dbContext.Coupons
                     .FirstOrDefaultAsync(c => c.Team.Id == team.Id && c.Id == model.CouponId);
             }
-
             // manage multiple customer scenario
             var invoices = new List<Invoice>();
             foreach (var customer in model.Customers)
@@ -102,6 +139,14 @@ namespace Payments.Mvc.Services
                 });
                 invoice.Attachments = attachments.ToList();
 
+                // start tracking for db
+                invoice.UpdateCalculatedValues();
+
+                if (invoice.CalculatedTotal <= 0)
+                {
+                    throw new ArgumentException("Invoice total must be greater than 0.", nameof(model));
+                }
+
                 if (model.Type == Invoice.InvoiceTypes.Recharge)
                 {
                     var rechargeAccounts = model.RechargeAccounts.Select(a => new RechargeAccount()
@@ -116,15 +161,19 @@ namespace Payments.Mvc.Services
 
                     });
                     invoice.RechargeAccounts = rechargeAccounts.ToList();
+
+                    if(invoice.CalculatedTotal != invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Credit).Sum(a => a.Amount))
+                    {
+                        throw new ArgumentException("Total of credit recharge accounts must equal invoice total.", nameof(model.RechargeAccounts));
+                    }
+                    if(invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Debit).Any() &&
+                        invoice.CalculatedTotal != invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Debit).Sum(a => a.Amount))
+                    {
+                        throw new ArgumentException("Total of debit recharge accounts must equal invoice total when supplied.", nameof(model.RechargeAccounts));
+                    }
                 }
 
-                // start tracking for db
-                invoice.UpdateCalculatedValues();
 
-                if (invoice.CalculatedTotal <= 0)
-                {
-                    throw new ArgumentException("Invoice total must be greater than 0.", nameof(model));
-                }
 
                 _dbContext.Invoices.Add(invoice);
 
@@ -139,6 +188,7 @@ namespace Payments.Mvc.Services
             // get team out of invoice
             var team = invoice.Team;
 
+            //TODO: Continue to do the account for now, but we may not need it for recharge invoices.
             // find account
             var account = await _dbContext.FinancialAccounts
                 .FirstOrDefaultAsync(a => a.Team.Id == team.Id && a.Id == model.AccountId);
@@ -191,8 +241,17 @@ namespace Payments.Mvc.Services
             });
             invoice.Items = items.ToList();
 
+            //TODO: Move this into a private method. Need to do on update too.
             if (invoice.Type == Invoice.InvoiceTypes.Recharge)
             {
+                // Recharges don't have tax
+                model.TaxPercent = 0;
+                invoice.TaxPercent = 0;
+                if(model.CouponId > 0)
+                {
+                    throw new ArgumentException("Coupons are not allowed for recharge invoices.", nameof(model.CouponId));
+                }
+
                 // remove old recharge accounts
                 // If we do it this way, there may be issues if an invoice is rejected, then we edit it and resend. Now all the entered by values are lost.
                 _dbContext.RechargeAccounts.RemoveRange(invoice.RechargeAccounts);
@@ -209,6 +268,32 @@ namespace Payments.Mvc.Services
 
                 });
                 invoice.RechargeAccounts = rechargeAccounts.ToList();
+
+                // Validate recharge accounts again.
+                if (invoice.RechargeAccounts == null || !invoice.RechargeAccounts.Any(a => a.Direction == RechargeAccount.CreditDebit.Credit))
+                {
+                    throw new ArgumentException("At least one credit recharge account is required for recharge invoices.", nameof(model.RechargeAccounts));
+                }
+
+                if (invoice.RechargeAccounts.Where(a => a.Amount <= 0.0m).Any())
+                {
+                    throw new ArgumentException("Recharge account amounts must be greater than 0.", nameof(model.RechargeAccounts));
+                }
+
+                foreach (var ra in invoice.RechargeAccounts)
+                {
+                    var validationModel = await _aggieEnterpriseService.IsRechargeAccountValid(ra.FinancialSegmentString, ra.Direction);
+                    if (!validationModel.IsValid)
+                    {
+                        throw new ArgumentException($"Recharge account '{ra.FinancialSegmentString}' is not valid");
+                    }
+                    // Ok, this could be called from an API, so we need to replace and chart strings that may get changed.
+                    if (validationModel.ChartString != ra.FinancialSegmentString)
+                    {
+                        ra.FinancialSegmentString = validationModel.ChartString;
+                        //log it?
+                    }
+                }
             }
 
             // add attachments
@@ -234,8 +319,23 @@ namespace Payments.Mvc.Services
                 throw new ArgumentException("Invoice total must be greater than 0.", nameof(model));
             }
 
+            if (invoice.Type == Invoice.InvoiceTypes.Recharge)
+            {
+
+                if (invoice.CalculatedTotal != invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Credit).Sum(a => a.Amount))
+                {
+                    throw new ArgumentException("Total of credit recharge accounts must equal invoice total.", nameof(model.RechargeAccounts));
+                }
+                if (invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Debit).Any() &&
+                    invoice.CalculatedTotal != invoice.RechargeAccounts.Where(a => a.Direction == RechargeAccount.CreditDebit.Debit).Sum(a => a.Amount))
+                {
+                    throw new ArgumentException("Total of debit recharge accounts must equal invoice total when supplied.", nameof(model.RechargeAccounts));
+                }
+            }
+
             return invoice;
         }
+
 
         public async Task SendInvoice(Invoice invoice, SendInvoiceModel model)
         {
@@ -243,6 +343,20 @@ namespace Payments.Mvc.Services
             if (!invoice.Sent)
             {
                 SetInvoiceKey(invoice);
+            }
+
+            //Possibly we want to validate the recharge accounts again before sending?
+            if(invoice.Type == Invoice.InvoiceTypes.Recharge)
+            {
+                foreach (var ra in invoice.RechargeAccounts)
+                {
+                    var validationModel = await _aggieEnterpriseService.IsRechargeAccountValid(ra.FinancialSegmentString, ra.Direction);
+                    if (!validationModel.IsValid)
+                    {
+                        //This will throw an uncontrolled exception, but that is probably better than sending an invoice with invalid accounts.
+                        throw new ArgumentException($"Recharge account '{ra.FinancialSegmentString}' is not valid");
+                    }
+                }
             }
 
             await _emailService.SendInvoice(invoice, model.ccEmails, model.bccEmails);
@@ -290,6 +404,18 @@ namespace Payments.Mvc.Services
 
             return linkId;
         }
+
+        public async Task SendFinancialApproverEmail(Invoice invoice, SendApprovalModel model)
+        {
+            if (invoice.Type != Invoice.InvoiceTypes.Recharge)
+            {
+                return;
+            }
+
+            //The cc emails are actually going to be the to emails in this case.
+            //We might want to change it a little to have the names as well.
+            await _emailService.SendFinancialApprove(invoice, model);
+        }
     }
 
     public interface IInvoiceService
@@ -299,6 +425,8 @@ namespace Payments.Mvc.Services
         Task<Invoice> UpdateInvoice(Invoice invoice, EditInvoiceModel model);
 
         Task SendInvoice(Invoice invoice, SendInvoiceModel model);
+
+        Task SendFinancialApproverEmail(Invoice invoice, SendApprovalModel model);
 
         Task RefundInvoice(Invoice invoice, PaymentEvent payment, string refundReason, User user);
